Financial Module Implemented
First-cut implementation of the SAM financial and resource modules. I
will do more work on this to make the command flow a little bit more
sensible as time goes by.  Right now there's a little too much linear
dependence that will need to be broken in order to allow the most
generic form of user inputs.